fake backend

+++ src/app/_services/support-user-data.service.ts
@@ -5,6 +5,8 @@ import { map, retry } from 'rxjs/operators';
 import { environment } from '../../environments/environment';
 import { UserAccessModel } from '../models/user-access-model';
 import { ValueProcessingService } from './value-processing.service';
+import { AuthService } from './auth.service';
+import { Role } from '../models/role';
 
 @Injectable({
   providedIn: 'root'
@@ -20,7 +22,7 @@ export class SupportUserDataService implements OnDestroy {
   httpOptionsNoKey: any;
   totalCountSubject = new BehaviorSubject<number>(0);
 
-  constructor(private http: HttpClient, private valueService: ValueProcessingService) {
+  constructor(private http: HttpClient, private auth: AuthService, private valueService: ValueProcessingService) {
     this.httpOptions = {
       headers: new HttpHeaders({
         // 'Ocp-Apim-Subscription-Key': environment.users.ocpApimSubscriptionKey,
@@ -44,6 +46,9 @@ export class SupportUserDataService implements OnDestroy {
   }
 
   public getUsers(params: any): Observable<UserAccessModel[]> {
+    if (this.auth.getUserRole() !== Role.Admin) {
+      return null;
+    }
     return this.http.post(this.postUrl, params, this.httpOptions)
       .pipe(map<any, UserAccessModel[]>(data => {
         this.totalCountSubject.next(data.totalCount);
@@ -53,16 +58,25 @@ export class SupportUserDataService implements OnDestroy {
   }
 
   getTotalCount(): Observable<number> {
+    if (this.auth.getUserRole() !== Role.Admin) {
+      return null;
+    }
     return this.totalCountSubject.asObservable();
   }
 
   getById(id: number): Observable<UserAccessModel> {
+    if (this.auth.getUserRole() !== Role.Admin) {
+      return null;
+    }
     return this.http.post(this.userUrl.replace('{user_id}', id.toString()), {}, this.httpOptions)
       .pipe(map<any, UserAccessModel>(data => data.items != null && data.items.length > 0 ? data.items[0] : null)
     );
   }
 
   createUser(user: UserAccessModel, successCallback, errorCallback) {
+    if (this.auth.getUserRole() !== Role.Admin) {
+      return null;
+    }
     const request = {
       user: {
           firstName: user.firstName,
@@ -91,6 +105,9 @@ export class SupportUserDataService implements OnDestroy {
   }
 
   update(user: UserAccessModel, successCallback, errorCallback) {
+    if (this.auth.getUserRole() !== Role.Admin) {
+      return null;
+    }
     const request = {
         id: user.id,
         firstName: user.firstName,
@@ -114,6 +131,9 @@ export class SupportUserDataService implements OnDestroy {
   }
 
   delete(userId: number, successCallback, errorCallback) {
+    if (this.auth.getUserRole() !== Role.Admin) {
+      return null;
+    }
     this.http.delete(this.userUrl.replace('{user_id}', userId.toString()), this.httpOptions)
       .pipe(
         retry(3)
